Documented the cluster algorithm used by Cluster.
While care was taken to add documentation and renaming variables to
improve understanding without changing the logic, it was modified in
one way: an unnecessary variable `done` was replaced by just checking
whether `accumulating_cluster` had been set or not. We should carefully
test to ensure that the clustering algorithm has not changed before
merging this code.